dashboard_app/api/requirements.txt
gunicorn==19.6.0
gevent==1.2.1
json-logging-py==0.2
flask-restful==0.3.6
smart-getenv==1.1.0
webargs==1.8.1
flask-cors==3.0.3
flask-migrate==2.1.1
flask-sqlalchemy==2.3.1
flask_script==2.0.6
psycopg2-binary==2.7.4
requests==2.18.4
flask_testing==0.7.1
faker==0.8.13